+++ CXXCore-master/include/cxxvector.h
@@ -1,10 +1,10 @@
 ﻿/*
 				GNU GENERAL PUBLIC LICENSE
-				   Version 2, June 1991
+				   Version 3, June 1991
 COPYRIGHT (C) Bestkakkoii 2024 All Rights Reserved.
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
+the Free Software Foundation; either version 3 of the License, or
 (at your option) any later version.
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of