+++ module_cxxcore.h
@@ -1,10 +1,10 @@
 ﻿/*
 				GNU GENERAL PUBLIC LICENSE
-				   Version 2, June 1991
+				   Version 3, June 1991
 COPYRIGHT (C) Bestkakkoii 2024 All Rights Reserved.
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
+the Free Software Foundation; either version 3 of the License, or
 (at your option) any later version.
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -60,8 +60,14 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 // - **炫文本默认编码**: UTF-16LE
 // 
 // # 最新更新日志:
-// ####[2024 - 10 - 06]()
-// * 修改 应对 IDE 总是出现命名空间解析 BUG，将所有命名空间的枚举移至核心模块
+// ####[2024-10-10]()
+// * 修改 将 `cxxmemorypool.h` 和 `cxxmemorypool.cpp` 的版权声明从 Box2D 版权声明更改为 GNU General Public License(GPL)。
+//   * 将 `cxxmemorypool.h` 中的 `#include <cxxglobal.h>` 和 `#include <cstdint>` 替换为 `#include <memory_resource>`。
+//	 * 移除了原本的内存池实现，并改用 C++17 的 `std::pmr::synchronized_pool_resource` 来实现内存池，并将内存池的默认大小设置为 1MB(`MemoryPoolDefaultPoolSize`)。
+//   * 更新了内存池的优点和缺点说明，并添加了使用方法的註释。
+//   * 将 `CXXMemoryPool` 类模板化，并修改了内存分配和释放的方法签名，使其支持分配和释放指定数量的对象。
+//   * 更新了 `CXXScopedMemoryPool` 类，使其与新的 `CXXMemoryPool` 类兼容。
+//   * 在 `CXXCoreTest.vcxproj.filters` 中，将 `cxxmemorypool` 的过滤器从 `old` 文件夹移动到 `template` 文件夹。
 //
 // [Gitee](https://gitee.com/Bestkakkoii/CXXCore)
 // [Github](https://github.com/bestkakkoii/CXXCore)