+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ﻿## 更新日志
 
+#### [2024-10-10]()
+  * 修改 将 `cxxmemorypool.h` 和 `cxxmemorypool.cpp` 的版权声明从 Box2D 版权声明更改为 GNU General Public License (GPL)。
+    * 将 `cxxmemorypool.h` 中的 `#include <cxxglobal.h>` 和 `#include <cstdint>` 替换为 `#include <memory_resource>`。
+    * 移除了原本的内存池实现，并改用 C++17 的 `std::pmr::synchronized_pool_resource` 来实现内存池，并将内存池的默认大小设置为 1MB (`MemoryPoolDefaultPoolSize`)。
+    * 更新了内存池的优点和缺点说明，并添加了使用方法的註释。
+    * 将 `CXXMemoryPool` 类模板化，并修改了内存分配和释放的方法签名，使其支持分配和释放指定数量的对象。
+    * 更新了 `CXXScopedMemoryPool` 类，使其与新的 `CXXMemoryPool` 类兼容。
+    * 在 `CXXCoreTest.vcxproj.filters` 中，将 `cxxmemorypool` 的过滤器从 `old` 文件夹移动到 `template` 文件夹。
+
 #### [2024-10-07]()
   * 修改 在 `cxxwintoast.h` 中新增 `CXXToastResultCallback` 的 `typedef` 和相关註释。
     * 在 `cxxwritelocker.h` 中将析构函数改为虚拟函数。